refactor(source): move to configurator
