+++ source/configurator/git/config.go
