+++ source/source.go
@@ -1,19 +1,13 @@
 package source
 
 import (
-	"context"
-
-	"github.com/alexfalkowski/konfig/source/folder"
-	"github.com/alexfalkowski/konfig/source/git"
+	"github.com/alexfalkowski/konfig/source/configurator"
+	"github.com/alexfalkowski/konfig/source/configurator/folder"
+	"github.com/alexfalkowski/konfig/source/configurator/git"
 )
 
-// Configurator for source.
-type Configurator interface {
-	GetConfig(ctx context.Context, app, ver, env, cluster, cmd string) ([]byte, error)
-}
-
 // NewConfigurator for source.
-func NewConfigurator(cfg *Config) Configurator {
+func NewConfigurator(cfg *Config) configurator.Configurator {
 	if cfg.IsGit() {
 		return git.NewConfigurator(&cfg.Git)
 	}
